roles added

+++ server/index.js
@@ -30,48 +30,49 @@ mongoose
 app.post("/signup", async (req, res) => {
 	const newUser = { ...req.body };
 	//missing data
-	if (!newUser.userName || !newUser.email || !newUser.password) {
+	if (!newUser.userName || !newUser.password) {
 		return res
 			.status(400)
-			.json({ error: "Datos Inválidos al crear nuevo usuario" });
+			.send({ error: "Datos Inválidos al crear nuevo usuario" });
 	}
 	//username in use
 	const userFoundByName = await User.find({ userName: newUser.userName });
 	if (userFoundByName.length > 0) {
-		res.status(400).json({ error: "El nombre de usuario ya existe." });
-	}
-	//email in use
-	const userFoundByEmail = await User.find({ email: newUser.email });
-	if (userFoundByEmail.length > 0) {
-		return res.status(400).json({ error: "El email ya existe." });
+		return res.status(400).send({ error: "El nombre de usuario ya existe." });
 	}
 
 	const salt = await bcrypt.genSalt(10);
 	const encodedPass = await bcrypt.hash(newUser.password, salt);
-
-	try {
-		const savedData = await new User({
-			...newUser,
-			password: encodedPass,
-		}).save();
-		res.status(201).json(savedData);
-	} catch (e) {
-		console.error(e);
-		res.status(400).json({ error: "Hubo un problema." });
-	}
+	User({
+		...newUser,
+		password: encodedPass,
+		role: "basic",
+	})
+		.save()
+		.then((result) => {
+			res.status(201).send({
+				uid: result._doc._id,
+				userName: result._doc.userName,
+				role: result._doc.role,
+			});
+		})
+		.catch((e) => {
+			console.log(e);
+			res.status(400).send({ error: "Hubo un problema" });
+		});
 });
 //signin
 app.post("/signin", async (req, res) => {
 	const userData = { ...req.body };
 	//missing data
 	if (!userData.userName || !userData.password) {
-		return res.status(400).json({ error: "Datos Inválidos al ingresar" });
+		return res.status(400).send({ error: "Datos Inválidos al ingresar" });
 	}
 	try {
 		const userFoundByName = await User.findOne({ userName: userData.userName });
 
 		if (!userFoundByName) {
-			return res.status(401).json({ error: "Usuario desconocido" });
+			return res.status(401).send({ error: "Usuario desconocido" });
 		}
 
 		const passOK = await bcrypt.compare(
@@ -82,19 +83,26 @@ app.post("/signin", async (req, res) => {
 			return res.status(401).json({ error: "Contraseña Inválida" });
 		}
 
-		res.status(201).json({ ...userFoundByName._doc });
+		res.status(201).send({
+			uid: userFoundByName._id,
+			userName: userFoundByName.userName,
+			role: userFoundByName.role,
+		});
 	} catch (e) {
 		console.error(e);
+		res.status(400).send({ error: "Hubo un problema" });
 	}
 });
-
 //Add Item
 app.post("/add", (req, res) => {
-	const item = new Item({
-		item: req.body.item,
-		cant: req.body.cant,
-	});
-	item
+	const { uid, item, cant } = req.body;
+
+	const itemObject = new Item({ uid, item, cant });
+	if (!uid || !item || !cant) {
+		return res.status(400).send({ error: "Datos Inválidos al crear item" });
+	}
+
+	itemObject
 		.save()
 		.then((result) => {
 			res.json({ error: null, ...result._doc });
@@ -106,19 +114,28 @@ app.post("/add", (req, res) => {
 });
 //Update Item
 app.post("/update", async (req, res) => {
-	const { _id, item, cant } = req.body;
-	await Item.updateOne({ _id }, { item, cant });
-	res.json({ error: null, message: _id + " Editado" });
+	const { _id, item, cant, uid } = req.body;
+	const user = await User.findById(uid);
+	if (user.role === "admin") {
+		await Item.updateOne({ _id }, { item, cant, uid });
+		res.status(200).send({ message: _id + " Editado" });
+	} else {
+		res.status(400).send({ error: "Solo los administradores pueden editar" });
+	}
 });
-
 //Delete Item
 app.post("/delete", async (req, res) => {
-	const { _id } = req.body;
-	await Item.deleteOne({ _id });
-	res.json({ error: null, message: _id + " Borrado" });
+	const _id = req.body.itemId;
+	const uid = req.body.uid;
+	const user = await User.findById(uid);
+	if (user.role === "admin") {
+		await Item.deleteOne({ _id });
+		res.status(200).send({ message: _id + " Borrado" });
+	} else {
+		res.status(400).send({ error: "Solo los administradores pueden eliminar" });
+	}
 });
-
-//Load data
+//Load Items data
 app.get("/load", (req, res) => {
 	Item.find()
 		.sort("item")
@@ -128,4 +145,25 @@ app.get("/load", (req, res) => {
 				data: result,
 			});
 		});
+});
+//Load Users data
+app.get("/user/load", (req, res) => {
+	User.find().then((result) => {
+		res.json({
+			error: null,
+			data: result,
+		});
+	});
+});
+
+//Update User
+app.post("/user/update", async (req, res) => {
+	const { _id, role, uid } = req.body;
+	const user = await User.findById(uid);
+	if (user.role === "admin") {
+		await User.updateOne({ _id }, { role, uid });
+		res.status(200).send({ message: _id + " Editado" });
+	} else {
+		res.status(400).send({ error: "Solo los administradores pueden editar" });
+	}
 });
